Tiga belas bug sisa di engine, juel, model-api dan webapp
S2175 (3) - pencarian di koleksi yang tipenya tidak mungkin cocok.
ProcessEngines menyimpan info dengan kunci berupa string tetapi mencarinya
dengan objek URL, sehingga blok pembersihan registrasi lama tidak pernah
jalan dan entri basi menumpuk; MigratingActivityInstanceVisitor mencari
peta ber-kunci PvmActivity memakai ScopeImpl; selectBoolean mendeklarasikan
List<String> padahal statement-nya resultType="integer".

S2259 (5) - deleteRoot dipakai di luar pemeriksaan null-nya sendiri; id
form field dibandingkan setelah addError walau null; sourceExpression
dibaca di bawah penjagaan linkExpression yang tidak menjaminnya; state
history job yang tidak dikenali tidak lagi dipanggil getStateCode();
restoreProcessInstance melempar pesan yang jelas ketika tidak ada eksekusi
process instance.

S1143 (1) - throw di dalam finally membuang kegagalan service.stop();
sekarang kegagalan itu ditahan dan dilampirkan sebagai suppressed.

S2153, S1226, S2177, S4351 - boxing Short yang langsung di-unbox, parameter
yang ditimpa, nama method privat yang bertabrakan dengan induknya, dan
compareTo(String) pada enum yang bukan pembanding urutan.

Diverifikasi: 359 test juel, 118 test xml-model, 731 test webapp, nol gagal.

https://collab.javan.co.id/projects/713/tasks/138524

+++ juel/src/main/java/org/camunda/bpm/impl/juel/TypeConverterImpl.java
@@ -241,7 +241,8 @@ public class TypeConverterImpl implements TypeConverter {
       }
     }
     if (value instanceof Character) {
-      return Short.valueOf((short) ((Character) value).charValue()).byteValue();
+      // boxing into a Short only to unbox it again as a byte; the cast does the same thing
+      return (byte) ((Character) value).charValue();
     }
     throw new ELException(LocalMessages.get("error.coerce.type", value.getClass(), Byte.class));
   }
